apps/python_api/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
alembic>=1.12.0
pydantic>=2.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
celery>=5.3.4
redis>=5.0.0
supabase>=1.0.3
pytest>=7.4.2
httpx>=0.25.0
starlette>=0.27.0
email-validator>=2.0.0
websockets>=11.0.3
requests>=2.31.0
psutil>=5.9.5
bcrypt>=4.1.2
